backend/requirements.txt
Django==6.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.1
Pillow==12.0.0
psycopg[binary]==3.3.2
python-decouple==3.8
drf-yasg==1.21.7
django-filter==25.2
gunicorn==21.2.0
